Remove duplicate LC_RPATH values on macOS
Starting with Qt 5.9, the values in QMAKE_RPATHDIR were passed along
to GN, which resulted in both qmake and GN adding rpath linker flags,
leading to duplicate rpath values.

This lead to the Qt package installer not removing all absolute path
rpath values when installing Qt, which subsequently caused issues
with code signing Qt applications.

The fix is not to pass the rpath values to GN, and rely on the ones
generated by qmake.

Task-number: QTBUG-61413
Change-Id: Ib27d9c232674534d4b053cb1e12750f5c8bc87a9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>