Windows: Make sure that the path is correctly quoted
We need to ensure that the path is using forward slashes to avoid any
chances of a trailing backslash escaping the final quote. Additionally
we need to quote the whole string in case there is a space, we can't use
shell_quote for this as this preserves the backslashes which can cause
a problem.

Change-Id: Iadde9a10e2d80ad48efd695d414af58d891c7412
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>