Document about popups not appearing on Windows when in fullscreen mode
On Windows there is an issue that an OpenGL-backed fullscreen window
will prevent other top-level windows from appearing on top of it (for
example a select / combo box popup). This appears to be a limitation of
the Windows platform compositor (DWM).
The workaround is to call
QWindowsWindowFunctions::setHasBorderInFullScreen on the QWindow which
will show the fullscreen WebEngine view.

Task-number: QTBUG-61563
Change-Id: I5e0f08ef49de6119ef7910cec2b32ea267301c18
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>