Add a QTWEBENGINE_DICTIONARIES_PATH environment variable
On a Linux system with an installed application, the only places dictionaries
are searched currently are /usr/bin/qtwebengine_dictionaries and
/usr/share/qt/qtwebengine_dictionaries (or similar, depending on the
distribution).

This adds an environment variable which makes it possible to customize the
location dictionaries are in (e.g. to something user-writable).

Documentation is added in the next commit, as it was duplicated in four
different locations and I cleaned it up a bit.

[ChangeLog] New QTWEBENGINE_DICTIONARIES_PATH environment variable to customize
the location for spell checking dictionaries.

Task-number: QTBUG-62236
Change-Id: Iefc90451c537656025abb785291fe55f6a82d50c
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
Reviewed-by: Viktor Engelmann <viktor.engelmann@qt.io>