Stop forwarding HoverEnter and HoverLeave QEvents
With QtWidgets the QHoverEvent handling is slightly incorrect
(HoverEnter and HoverLeave triggering Q_ASSERT), quite unnecessary
(Chromium works fine with just MouseMove events when mouse tracking is
enabled), and mostly unused (QHoverEvents are only delivered if the
WA_Hover widget attribute is set, which it usually is not).

QtQuick however does not have the equivalent of QtWidgets mouse
tracking, so to get mouse movement information into Chromium we have to
use HoverMove QEvents. But the HoverEnter and HoverLeave QEvents are not
used or useful for QtQuick either.

Task-number: QTBUG-62200
Change-Id: I333de2b6adcc24544935d36645036aedb07e51ac
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>