Close popup windows when their parents are destroyed
Previously when a popup window was shown, and the user issued a command
to close the only active main window (via a shortcut for example) with
a subsequent call to deleteLater(), the main window would be closed
and deleted but the popup would only be hidden, which led to the
application loop to still be running and not quitting, even though
there would be no visible window left.

The closing of a popup is usually done in the QWidget destructor when
the popup is still visible, but because we unset the parent right
before the parent is destroyed (which sets visibility to false), the
popup would not be closed, but only hidden.

Thus this change makes sure to explicitly close the popup when its
parent is destroyed.

Task-number: QTBUG-62311
Change-Id: Ia0bbf327929af55aac19767971ab0acde5715e21
Reviewed-by: Viktor Engelmann <viktor.engelmann@qt.io>