Fix position of keyboard-invoked context menu in QQuickWebEngineView
In QQuickWebEngineView the context menu is a QtQuickControls.Menu item.
This menu is shown by calling popup() which always displays the menu
below the mouse cursor.

Work around the problem by moving the mouse cursor temporarily to the
right position.

Use a QObject property "pos" to store the requested menu position
between addMenu() and showMenu() calls, because the Menu item doesn't
have a "pos" QML property.

Change-Id: Id772a0bb1a7548cad932e9f499ade68be32d86d3
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>