Document that the Windows version in the user-agent might be wrong
Chromium uses GetVersionEx() to determine the Windows version that is
reported in the default user agent. GetVersionEx() however always
reports "Windows NT 6.2" for newer Windows versions, unless the
executable's manifest file marks the newer version as supported.

Since this is a common question we document this in the httpUserAgent
accessors.

Task-number: QTBUG-56472
Change-Id: I4698cd659f5552b92ef925c198f39326fcb936f3
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>