[macOS] Fix creation of shared OpenGL 3.2 Core contexts
Due to a current issue described in QTBUG-63180, requesting a 3.2 Core
context on macOS will advertise that a 4.1 context was created, even
though that is not the case. Because
RenderWidgetHostViewQtDelegateWidget will read the OpenGL major version
from the global shared context and then apply that to its own context,
this will cause a failure to create a shared context (one is 3.2, and
the other is 4.1).

The current workaround is to create the context with the default format
version, instead of the global shared one. This way all the requested
versions will be consitent.

Task-number: QTBUG-60605
Change-Id: I470c43ca9d15cb3887a0ed968b57c62518a33a72
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>