Make GetUserMediaTestPage more robust
A 5 second timeout might not be enough for our underpowered CI VMs.
Also, if the page isn't loaded (to slow or for other reasons), we
must abort, to prevent the JavaScripts from being run on different
pages (first scripts to the old page, then to the new page if it
eventually finishes loading).
The first JavaScript in particular is rejected on the old page,
because "about:blank" doesn't have a secure origin. This could
even make the test accept wrongfully if all scripts are run on the
old page, because the request (which is supposed to be rejected)
can be rejected for being run on the wrong page, instead of for
requesting forbidden permissions.

Task-number: QTBUG-63347
Change-Id: Ib63d55477d35001b174900b73dc3dd23d33c7263
Reviewed-by: Jüri Valdmann <juri.valdmann@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>