Fix viewport resizing bug in renderer
After the recent fix for rendering intersecting quads, DelegatedFrameNode no
longer builds scene graph nodes for DrawQuads that are outside the visible area.
This means that the structure of the scene graph now depends on the size of the
visible area, however the logic for deciding whether to update or rebuild the
scene graph was not updated to reflect this fact. As a result we may try to
update e.g. a QSGImageNode as if it were a QSGRectangleNode leading to a crash.

Task-number: QTBUG-62112
Change-Id: I6e2e9dee4238d208fc2be98669281c2d4d4962d7
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>