Stabilize navigation after render process termination
When renderProcessTerminated signal is emitted the invalid view of the
corresponding RenderWidgetHost is not destructed yet. Having this
invalid view during navigation leads to a crash.

The proper way to navigate on renderProcessTerminated is to schedule the
navigation with QTimer::singleShot(). Otherwise, warn the user and ignore
navigation to avoid crash.

Task-number: QTBUG-58478
Change-Id: I54ff62f5d306c8cb1fa17e29f349d1e02c4c8cec
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>