Fix qwebengine_convert_dict to pick up correct ICU resources location
The conversion tool uses the QLibraryInfo::location() call to find Qt's
data directory. When a qt.conf file gets picked up by the application
(which is the case for official Qt packages), QLibraryInfo::location()
depends on a QCoreApplication instance being available, to query the
PrefixPath. If the instance is not available, it simply returns the
current directory path, which results in ICU resources not being
found.

The fix consists in creating a QCoreApplication instance.

Task-number: QTBUG-63593
Change-Id: I8f1bba14d4586c42570d309b84418eef611bab0e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>