Process pending resize of view before scroll to anchor by URL fragment
The size of the WebEngineView must be set before navigating to an URL
with fragment for scrolling to the correct position. In some cases the
resizing is scheduled on the event loop and Blink uses wrong size to
calculate position to scroll. Schedule the navigation on the event loop
if the viewport's size is 0 thus it is performed after the potential
resizing is done.

Task-number: QTBUG-54172
Change-Id: I9acbc33748755606082278585c121dd66b4aa02c
Reviewed-by: Viktor Engelmann <viktor.engelmann@qt.io>
(cherry picked from commit 15b6ab6a20ff7305f2b3558f8450e388148a0312)
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>