Speculative stabilization of mouseLeave auto tests
Occasionally the mouse pointer might be moved above the testDiv during
the html load and this changes its text. This seems to happen on CI
from time to time.

Text of testDiv is expected to be empty in the beginning to make sure
an explicit mouse move event changes its text.

Change-Id: Ib8488b1e2dd7cc80b78cc40adecbf04c12dbec1b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>