Hardcode default argument for mimetype in ::setContent
The docs were wrong in that the default is actually text/plain with
US-ASCII encoding (see e.g. data_url_unittest.cc).

Anyhow, saying it might change in the future does not help anyone,
and is actually a potential security risk - see e.g. the discussion
about 'magic' in QTextEdit::setText(). Because of this, it's
unlikely that Chromium ever changes the default. Anyhow, we can as
well hardcode the default, and document it then.

Change-Id: I949111598a30fa69d152d3e98d76e9d96df92d54
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Viktor Engelmann <viktor.engelmann@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>