Turn off caching of images rendered for the printer
When printing very large documents using the QPrinter-based API, it
was possible to run out of memory because the images that pdfium has
generated was cached for reuse when printing multiple copies of the
document.

Caching the images is now removed as printing multiple copies is not
the default use case and is nowadays often also handled by the printer
itself.

Task-number: QTBUG-58400
Change-Id: I27bd17b33a839a845ca1b387b0c3bd0466b6592f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>