Prefix icu library in configure.json to prevent build issues on Win32
Both qtbase/src/corelib/configure.json and qtwebengine/configure.json
declare the "icu" library, but the definitions of the library are
different. Make sure to prefix the webengine "icu" declaration, so that
it does not override the qtbase one.

This is a partial backport of commit
2ed1e0ba423a75f68c1547fe0c06e62e773ef64d which was introduced in
Qt 5.10.0.

The rest of the library definitions are not conflicting at the moment.

Task-number: QTBUG-62944
Change-Id: If5cbf07a5bab149c7663ce05c7eb8183c5f63300
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>