Use correct margins when printing with QPrinter
Margins were applied two times: PrintViewManagerQt included them when
rendering to a temporary PDF, then QPrinter applied them when printing.

This patch changes the parameters of PDF generation to not contain the
area of margins. This way the user-provided QPrinter can handle it
without any kind of customization.

Task-number: QTBUG-65715
Change-Id: Ie32785857e5195aa21c6e2ccb7ebd1e6f6ecd954
Reviewed-by: Michael Brüning <michael.bruning@qt.io>