Prevent Qt Quick Compiler from removing qrc Javascript source code
By default whenever a Qt application is built with Qt Quick compiler,
it will scan all qrc resources for javascript files, generate
corresponding C++ code, and remove the original JS code. This is
problematic for web pages that are loaded in WebEngine because the pages
won't be able to access the javascript code, and silently fail (or if
you are lucky, you will get some unrelated "can not be found" error).

Use the undocumented QtQuick Compiler
QTQUICK_COMPILER_SKIPPED_RESOURCES qmake variable to specify that
Javascript source code found in qrc files should be left as is.

Task-number: QTBUG-66155
Change-Id: Ib1dc744901baddd9e0c17c53db1ffede5c31e544
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>