Consider various resources fallback paths on macOS framework build
When building a macOS framework build, paths like process path,
resources path, icu path, etc are determined solely by the existence
of the WebEngineCore.framework bundle, which is found using its bundle
id.

There might be cases when the bundle is not present. For example when
deploying a WebEngine PySide2 application using PyInstaller, the layout
of the copied files is changed.

In this case considering application path (current application
directory) as a candidate path, as well as the rest of the candidates
paths would be useful.

Change-Id: Ide3fd62756659ec8d42c5629c0efa1d60602ed80
Reviewed-by: Michael Brüning <michael.bruning@qt.io>