Add support for loading UI delegates from resource files
Custom UI delegates can be specified for dialogs, context menus, etc. by
placing them in a specific folder accessible from the import path, but
it assumes that all import paths are folders on the os filesystem.

But resource (.qrc) files are also supported as a built-in import path,
so we should support loading UI delegates from there too.
We just need a couple of adjustments for QFileInfo to look them up using
the ":/" prefix, while still loading then using the "qrc:/" protocol.

Also add support for partial controls override, so it's possible
to customize only some UI delegates by prepending an import path with
some delegate files in there, and use the built-in ones for the rest.

Change-Id: I6c791ed0118b396639acd8af8e135e1d68b2c19b
Reviewed-by: Michal Klocek <michal.klocek@qt.io>