Apply settings on pending RVH too
This is an improvement of:
    bab4abab Keep settings synchronized when RVH swaps

Updating settings during RVH swap is flaky because it happens after
the load start of RenderFrame thus it is not guaranteed the
ViewMsg_UpdateWebPreferences message is received by the render process
before the load finishes. This means that the setting might not be applied
on the current page if it was set during the navigation.

The new RVH is created at the very beginning of the navigation, so access
it via the pending or speculative main frame of the page thereby make it
possible to update its settings before the load of RenderFrame starts.

Task-number: QTBUG-66656
Task-number: QTBUG-68424
Change-Id: I4fbba597579551bb3329936ccd7b357cf8daa4ea
Reviewed-by: Jüri Valdmann <juri.valdmann@qt.io>