Fix bogus check for GL_ARB_ES2_compatibility
We check for GL_ARB_ES2_compatibility on wayland when
shared context is created for opengl and we want
to check if es2 compatibility is possible.
However check should be done on OpenGL renderableType
surface and not the other way around. This was
currently working due to bug in qtwayland, which
returned here OpenGL context instead of requested OpenGLES.

Change-Id: I47ab02cb9128670bfb02096d5911c4dcd24284d2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>