Leave the decision of including RTTI information to the compiler
This change removes the Chromium default of stripping RTTI information
from our own cpp files in QtWebEngineCore. This means RTTI information
will be emitted based on the compiler's defaults, which follows the
same principle as qtbase, which does not explicitly include or exclude
the information.

A user can still override this by adding CONFIG+=rtti (or rtti_off),
which will propagate correctly to GN (but this only really applies
to Windows).

DLL size difference on a debug build is only 78KiloBytes.

Usecase is that RTTI info is currently required by PySide2 for
marshalling various QObjects to and from Python, and identifying
dynamic types by means of C++'s typeid().

Change-Id: Idb9dacbd1760dbf5b29736b2f2c731b9f6514885
Fixes: PYSIDE-819
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>