Lock profile mutex during interceptRequest
This makes setRequestInterceptor take effect atomically, so it's always safe to
delete the interceptor after unsetting it from the profile. On the other hand,
we run into a deadlock if the user code tries to wait for a profile operation on
the UI thread to finish. It seems quite unlikely though that user code would do
that and it would probably still be easier for users to understand and debug
than the current non-deterministic use-after-free problems.

Task-number: QTBUG-72260
Change-Id: I1784a9d9f00029d440508f0bb076d1081a326758
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>