tst_origins: Don't create SharedWorkers from unregistered scheme
Triggers DCHECK in content::SharedWorkerInstance::Matches on CI. According to
the DCHECK, SharedWorkers cannot be constructed from unique origins (worker and
constructor page need to be same-origin and unique origins are never same-origin
with anything). In release builds the construction will just fail as expected.

The DCHECK is only triggered if at least one other active SharedWorker already
exists in the same storage partition. The reason it only happens on CI might be
that the previous renderer process is not torn down fast enough and the previous
SharedWorker remains active.

Change-Id: I0906e1b5c85587df67b768ef022386d599ce9541
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>