Fix Sandbox V2 to work on macOS
The renderer process crashed on macOS because sandboxing
was not enabled.

The patch does two things to fix that:
1) Initializes the Seatbelt Sandbox in processMain().
2) Exposes a content::getQtPrefix() function that will be used by
SetupCommonSandboxParameters in sandbox_parameters_mac.mm to pass the
Qt Prefix path to the sandbox rules in
chromium/services/service_manager/sandbox/mac/common_v2.sb
The path will be used to allow file read access (required to read ICU
data files and other resource files).

To check which resources are denied access, start any QtWebEngine app
with the "--enable-sandbox-logging" flag, and use Console.app to
see the denied requests.

Change-Id: I61c0a2d20d5e96c7cc9562fd203fd606dd25d36c
Fixes: QTBUG-73089
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>